Explaining how to run zef from the container
To tell you the truth, I had to discover it myself... It's easier if
you create a Dockerfile using this one as a base, since you can just
`RUN` `zef` inside it. Not so intuitive if you want to do it by hand.

Hope this closes #1, thanks to @moritz and @loltimo for the
suggestion. If someone can test if this helps for perl6/doc#1384 that
would be great.